Moved helpers to challenge library
